Rename water quality service to avoid conflict
